refactor: move modules to utils directory
