+++ src/main.ts
@@ -9,12 +9,12 @@ import momentComment from "./alpine-data/moment-comment";
 import "./assets/fonts/meslo-LG/MesloLGS-Regular.woff2";
 import "./assets/fonts/meslo-LG/MesloLGS-Regular.woff";
 import "./assets/fonts/meslo-LG/MesloLGS-Regular.ttf";
-import "./assets/animation.ts";
-import "./assets/main.ts";
+import "./utils/animation";
+import "./utils/main";
 import "./styles/main.css";
 import "./styles/tailwind.css";
 
-export * from "./create-toc";
+export * from "./utils/create-toc";
 
 window.Alpine = Alpine;
 Alpine.data("momentComment", momentComment);
